Refactorizado GameController en dos controladores que separan juego y lobby
